requirements.txt
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fastjsonschema==2.21.2
ipykernel==7.1.0
ipython==9.8.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.9.1
matplotlib-inline==0.2.1
narwhals==2.13.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
prompt-toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
warp-lang==1.10.1
wcwidth==0.2.14
